#60: refactor: 스타일 관련 상수값과 타입 파일 위치 분리
