+++ src/constants/styles/animations.ts
