+++ src/constants/styles/background.ts
