+++ src/constants/styles/colors.ts
@@ -1,0 +1,32 @@
+export const colors = {
+  yellow100: 'var( --color-yellow-100)',
+  yellow200: 'var( --color-yellow-200)',
+  orange100: 'var( --color-orange-100)',
+  orange200: 'var( --color-orange-200)',
+  orange300: 'var( --color-orange-300)',
+
+  pink100: 'var( --color-pink-100)',
+
+  green100: 'var( --color-green-100)',
+  green200: 'var( --color-green-200)',
+  green300: 'var( --color-green-300)',
+  green400: 'var( --color-green-400)',
+
+  blue100: 'var(--color-blue-100)',
+  blue200: 'var(--color-blue-200)',
+  blue300: 'var(--color-blue-300)',
+  blue400: 'var(--color-blue-400)',
+
+  gray100: 'var(--color-gray-100)',
+}
+
+export const shadows = {
+  orange100: 'var( --shadow-orange-100)',
+  pink100: 'var( --shadow-pink-100)',
+  pink200: 'var( --shadow-pink-200)',
+  blue100: 'var( --shadow-blue-100)',
+}
+
+export const gradients = {
+  yellow100: 'var(--gradient-yellow-100)',
+}
