+++ src/constants/styles/logo.ts
