+++ src/constants/styles/pointBtn.ts
