+++ src/constants/styles/triangle.ts
