+++ src/styles/colors.ts
@@ -1,3 +1,4 @@
+import { colors, shadows, gradients } from '@/constants'
 import { css } from '@emotion/react'
 
 export const colorPalette = css`
@@ -33,42 +34,3 @@ export const colorPalette = css`
     --gradient-yellow-100: rgba(115, 216, 156, 0);
   }
 `
-
-export const colors = {
-  yellow100: 'var( --color-yellow-100)',
-  yellow200: 'var( --color-yellow-200)',
-  orange100: 'var( --color-orange-100)',
-  orange200: 'var( --color-orange-200)',
-  orange300: 'var( --color-orange-300)',
-
-  pink100: 'var( --color-pink-100)',
-
-  green100: 'var( --color-green-100)',
-  green200: 'var( --color-green-200)',
-  green300: 'var( --color-green-300)',
-  green400: 'var( --color-green-400)',
-
-  blue100: 'var(--color-blue-100)',
-  blue200: 'var(--color-blue-200)',
-  blue300: 'var(--color-blue-300)',
-  blue400: 'var(--color-blue-400)',
-
-  gray100: 'var(--color-gray-100)',
-}
-
-const shadows = {
-  orange100: 'var( --shadow-orange-100)',
-  pink100: 'var( --shadow-pink-100)',
-  pink200: 'var( --shadow-pink-200)',
-  blue100: 'var( --shadow-blue-100)',
-}
-
-const gradients = {
-  yellow100: 'var(--gradient-yellow-100)',
-}
-
-export type Colors = keyof typeof colors
-
-export type Shadows = keyof typeof shadows
-
-export type Gradients = keyof typeof gradients
