+++ src/types/styles/background.ts
