--- src/types/styles/index.ts
@@ -1,4 +1,0 @@
-import { IMAGE_ADDRESS } from '@/constants'
-
-export type IMAGE_ADDRESS_KEY =
-  (typeof IMAGE_ADDRESS)[keyof typeof IMAGE_ADDRESS]
