+++ src/types/styles/logo.ts
