+++ src/types/styles/triangle.ts
